Redesign toggle input groups
